+++ packages/core/botpress/src/cli/create.js
@@ -54,7 +54,6 @@ const assertDoesntExist = file => {
  */
 const loadTemplate = async name => {
   const templatePath = path.join(__dirname, 'cli/templates/' + name)
-  const infoFile = path.join(templatePath, 'info.json')
 
   if (!fs.existsSync) {
     console.log(templateNotFoundError(name))
@@ -105,7 +104,7 @@ module.exports = () => {
     properties: {
       name: {
         description: chalk.white('module name:'),
-        pattern: /^[@botpress/|botpress-][a-z0-9][a-z0-9-_\.]+$/,
+        pattern: /^(\@botpress\/|botpress-)[a-z0-9][a-z0-9-_\.]+$/,
         message: `Name must be only lowercase letters, digits, dashes, underscores and dots.
 It must also start with "@botpress/" or "botpress-"`,
         required: true,