+++ packages/core/botpress/src/cli/start.js
@@ -61,8 +61,7 @@ module.exports = (projectPath, options) => {
     // eslint-disable-next-line no-eval
     const bf = eval('require')(botfile)
     const dataDir = util.getDataLocation(bf.dataDir, projectPath)
-    const modulesConfigDir = util.getDataLocation(bf.modulesConfigDir, projectPath)
-    return [dataDir, modulesConfigDir, 'node_modules']
+    return [dataDir, 'node_modules']
   }
 
   const opts = options.opts()
@@ -92,7 +91,7 @@ module.exports = (projectPath, options) => {
       setTimeout(() => process.exit(), 100)
     })
   } else {
-    const bot = new Botpress({ botfile })
+    const bot = new Botpress({ botfile, options })
     bot.start()
   }
 }