+++ packages/core/botpress/src/web/views/FlowBuilder/nodeProps/SkillCallNode.jsx
@@ -2,6 +2,7 @@ import React, { Component, Fragment } from 'react'
 
 import { Panel, Tabs, Tab, Badge, Button } from 'react-bootstrap'
 
+import PermissionsChecker from '~/components/Layout/PermissionsChecker'
 import EditableInput from '../common/EditableInput'
 import TransitionSection from './TransitionSection'
 
@@ -19,7 +20,7 @@ export default class SkillCallNodePropertiesPanel extends Component {
   }
 
   render() {
-    const { node } = this.props
+    const { node, readOnly } = this.props
 
     const onNameMounted = input => {
       if (input.value.startsWith('node-')) {
@@ -35,6 +36,7 @@ export default class SkillCallNodePropertiesPanel extends Component {
       <div className={style.node}>
         <Panel>
           <EditableInput
+            readOnly={readOnly}
             onMount={onNameMounted}
             value={node.name}
             className={style.name}
@@ -42,7 +44,9 @@ export default class SkillCallNodePropertiesPanel extends Component {
             transform={this.transformText}
           />
           <div style={{ padding: '5px' }}>
-            <Button onClick={editSkill}>Edit skill</Button>
+            <PermissionsChecker user={this.props.user} op="write" res="bot.skills">
+              <Button onClick={editSkill}>Edit skill</Button>
+            </PermissionsChecker>
             <Button onClick={seeFlow}>See flow</Button>
           </div>
         </Panel>
@@ -56,6 +60,7 @@ export default class SkillCallNodePropertiesPanel extends Component {
             }
           >
             <TransitionSection
+              readOnly={readOnly}
               items={node.next}
               header="Transitions"
               subflows={this.props.subflows}