+++ packages/dev-bot/botfile.js
@@ -1,4 +1,3 @@
-const isProd = process.env.NODE_ENV === 'production'
 const port = process.env.BOTPRESS_PORT || process.env.PORT || 3000
 const botUrl = process.env.BOTPRESS_URL || `http://localhost:${port}`
 
@@ -9,7 +8,7 @@ module.exports = {
   botUrl: botUrl,
 
   /*
-    The botpress environment, useful to disambiguate multiple 
+    The botpress environment, useful to disambiguate multiple
     instances of the same bot running in different environments.
     e.g. "dev", "staging", "production"
    */
@@ -52,12 +51,11 @@ module.exports = {
   mediaDir: './generated/media',
 
   /*
-    By default logs are enabled and available in `dataDir`
+    By default logs are enabled and stored in the DB for 30 days
    */
-  disableFileLogs: false,
-  log: {
-    file: 'bot.log',
-    maxSize: 1e6 // 1mb
+  logs: {
+    enabled: true,
+    keepDays: 30
   },
 
   /*
@@ -100,7 +98,7 @@ module.exports = {
     Access control of admin panel
   */
   login: {
-    enabled: process.env.NODE_ENV === 'production',
+    enabled: process.env.BOTPRESS_AUTH_ENABLED,
     useCloud: process.env.BOTPRESS_CLOUD_ENABLED,
     tokenExpiry: '6 hours',
     password: process.env.BOTPRESS_PASSWORD || 'password',