+++ packages/skills/botpress-skill-choice/CHANGELOG.md
@@ -3,6 +3,38 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+<a name="10.20.1"></a>
+## [10.20.1](https://github.com/botpress/modules/compare/v10.20.0...v10.20.1) (2018-06-21)
+
+
+
+
+**Note:** Version bump only for package @botpress/skill-choice
+
+<a name="10.20.0"></a>
+# [10.20.0](https://github.com/botpress/modules/compare/v10.19.0...v10.20.0) (2018-06-20)
+
+
+
+
+**Note:** Version bump only for package @botpress/skill-choice
+
+<a name="10.19.0"></a>
+# [10.19.0](https://github.com/botpress/modules/compare/v10.18.0...v10.19.0) (2018-06-19)
+
+
+
+
+**Note:** Version bump only for package @botpress/skill-choice
+
+<a name="10.18.0"></a>
+# [10.18.0](https://github.com/botpress/modules/compare/v10.17.3...v10.18.0) (2018-06-19)
+
+
+
+
+**Note:** Version bump only for package @botpress/skill-choice
+
 <a name="10.17.3"></a>
 ## [10.17.3](https://github.com/botpress/modules/compare/v10.17.2...v10.17.3) (2018-06-15)